Fixed global styles and footer is now legit sticky to the bottom
