+++ src/components/Footer.js
@@ -47,11 +47,8 @@ const Footer = ({ user }) => {
 export default Footer;
 
 const FooterWrap = styled.footer`
+  flex-shrink: 0;
   display: flex;
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  left: 0;
   justify-content: space-between;
   background-color: #636578;
   z-index: 10;
